chote: move `CitiesAPIClient` to production
